Disable SSR for the renderer app shell
